fixes for warnings

+++ cheb/init.c
@@ -64,7 +64,7 @@ void gsl_cheb_free(gsl_cheb_series * cs)
 int gsl_cheb_init(gsl_cheb_series * cs, const gsl_function *func,
                   const double a, const double b)
 {
-  int k, j;
+  size_t k, j;
 
   if(a >= b) {
     GSL_ERROR_VAL("null function interval [a,b]", GSL_EDOM, 0);
